drbd: convert conn_printk() to a macro
As it was, it would frequently print leading part and trailing payload
on different printk invokations, producing hard to read output on a busy
system (printk spread over two lines).

Also, using the macro, we get useful format mismatch warnings, so this
patch fixes those, too. Well, the only such warning my compiler produced.

+++ drbd/drbd_receiver.c
@@ -633,7 +633,7 @@ static int drbd_recv_all_warn(struct drbd_tconn *tconn, void *buf, size_t size)
 
 	err = drbd_recv_all(tconn, buf, size);
 	if (err && !signal_pending(current))
-		conn_warn(tconn, "short read (expected size %d)\n", size);
+		conn_warn(tconn, "short read (expected size %d)\n", (int)size);
 	return err;
 }
